refactor: split endpoints into routers
